+++ include/SZ3/utils/QuantOptimizatioin.hpp
@@ -123,7 +123,7 @@ int optimize_quant_invl_3d(const T *data, size_t r1, size_t r2, size_t r3, doubl
             freq_index = static_cast<ptrdiff_t>(mean_diff / precision) - 1 + radius;
         if (freq_index <= 0) {
             freq_intervals[0]++;
-        } else if (freq_index >= freq_intervals.size()) {
+        } else if (freq_index >= static_cast<ptrdiff_t>(freq_intervals.size())) {
             freq_intervals[freq_intervals.size() - 1]++;
         } else {
             freq_intervals[freq_index]++;
